extract message
