+++ src/components/Panel/style.tsx
@@ -1,4 +1,5 @@
-import { smallMobileAndUnder } from "@/constants";
+import { red500, smallMobileAndUnder } from "@/constants";
+import { addOpacityToHsla } from "@/helpers";
 import NextLink from "next/link";
 import styled from "styled-components";
 
@@ -40,4 +41,19 @@ export const SectionTitle = styled.h2`
   span {
     font-weight: 400;
   }
+`;
+
+const errorBackgroundColor = addOpacityToHsla(red500, 0.05);
+
+export const ErrorWrapper = styled.div`
+  width: min(100%, 512px);
+  min-height: 48px;
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  margin-top: 20px;
+  padding-inline: 16px;
+  background: ${errorBackgroundColor};
+  border: 1px solid var(--red-500);
+  border-radius: 2px;
 `;
